[update] compile each function to the trait Decode
